+++ src/Shared/IServerCertificateValidationEventSource.cs
@@ -1,0 +1,17 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+using System;
+
+namespace OpenTelemetry.ResourceDetectors;
+
+internal interface IServerCertificateValidationEventSource
+{
+    public void FailedToValidateCertificate(string error);
+
+    public void FailedToCreateHttpHandler(Exception exception);
+
+    public void CertificateFileDoesNotExist(string filename);
+
+    public void FailedToLoadCertificateInTrustedStorage(string filename);
+}