+++ src/OpenTelemetry.Exporter.Geneva/CHANGELOG.md
@@ -6,6 +6,9 @@
   is disposed.
   ([#1537](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1537))
 
+* Update OpenTelemetry SDK version to `1.8.0`.
+  ([#1635](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1635))
+
 ## 1.7.0
 
 Released 2023-Dec-11